requirements_full.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
redis==5.1.0
arq==0.26.1
anthropic==0.40.0
instructor>=0.6.0
google-genai>=1.0.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
pdfplumber==0.11.4
pymupdf==1.24.11
boto3==1.35.42
pydantic-settings==2.5.2
pydantic==2.9.2
PyYAML>=6.0
filetype==1.2.0
python-multipart>=0.0.22
sse-starlette==2.1.3
httpx==0.27.2
eml-parser==1.17.5
extract-msg==0.52.0
pgvector==0.3.5
cryptography>=46.0.6
streamlit>=1.54.0
streamlit-extras>=0.4.0
plotly==5.24.1
locust==2.32.1
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.2
fakeredis==2.26.1
respx==0.21.1
aiosqlite==0.20.0
ruff==0.6.9
mypy==1.11.2
pytest-cov==5.0.0
mcp>=1.0
rank-bm25>=0.2.2
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-exporter-prometheus>=0.48b0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
prometheus-client>=0.21.0
langsmith>=0.1.0
langfuse>=4.2.0
deepeval>=3.0.0
ragas>=0.4.3
langchain-anthropic>=0.3.0
datasets>=2.14.0
opentelemetry-instrumentation-anthropic>=0.48b0
structlog>=24.0.0

# Experiment tracking (optional — set WANDB_API_KEY to enable)
wandb>=0.16.0
fpdf2>=2.7.0
